Improve analytics bootstrapping
- Configure support for code-splitting
- Extracts analytics code out into a separate code-split module
- Configures analytics to use beacon transport where supported
- Configures maxScrollDepth tracker plugin
