+++ views/includes/metaFooterJS.njk
@@ -1,10 +1,3 @@
-<script id="script-app-config">
-    {# allows us to disable google analytics on live env when loaded via EC2/ELB and not the real domain #}
-    {# eg to avoid polluting data with extra domains #}
-    window._BLF = {
-        blockAnalytics: ({{ appData.environment == 'production' }} && window.location.hostname !== "{{ appData.config.get('siteDomain') }}")
-    };
-</script>
 <script src="//www.google-analytics.com/cx/api.js"></script>
 <script src="{{ 'javascripts/app.js' | getCachebustedPath }}"
         id="js-script-main"
