+++ views/includes/metaHeadJS.njk
@@ -7,22 +7,39 @@
             'addEventListener' in window &&
             'localStorage' in window &&
             'bind' in Function,
-        isOldIE: navigator.userAgent.indexOf('MSIE') !== -1 || navigator.appVersion.indexOf('Trident/') > 0,
+        isOldIE:
+            navigator.userAgent.indexOf('MSIE') !== -1 ||
+            navigator.appVersion.indexOf('Trident/') > 0,
         hasSeenPageBefore: (function() {
             var hasSeenPageBefore = false;
             try {
-                hasSeenPageBefore = !!window.sessionStorage.getItem('app.hasSeenPageBefore');
+                hasSeenPageBefore = !!window.sessionStorage.getItem(
+                    'app.hasSeenPageBefore'
+                );
             } catch (e) {} // eslint-disable-line no-empty
 
             if (!hasSeenPageBefore) {
                 try {
-                    window.sessionStorage.setItem('app.hasSeenPageBefore', true);
+                    window.sessionStorage.setItem(
+                        'app.hasSeenPageBefore',
+                        true
+                    );
                 } catch (e) {} // eslint-disable-line no-empty
             }
 
             return hasSeenPageBefore;
-        })()
+        })(),
+        /**
+         * Allows us to disable google analytics on live env when loaded via EC2/ELB
+         * and not the real domain e.g. to avoid polluting data with extra domains
+         */
+        blockAnalytics:
+            {{ appData.environment == 'production' }} &&
+            window.location.hostname !==
+                "{{ appData.config.get('siteDomain') }}"
     };
+
+    window.AppConfig = AppConfig;
 
     /**
      * Add state classes to document
@@ -45,4 +62,5 @@
 
     docEl.className = docClass.replace('no-js', '') + ' ' + classes.join(' ');
 })();
+
 </script>
